edge/requirements.txt
uvicorn==0.34.0
fastapi==0.115.6
pydantic==2.10.5
pika==1.3.2
tensorflow==2.17.0
pandas==2.2.2
scikit-learn==1.6.1
scipy==1.15.1
websockets==10.4.0
aio-pika==9.5.4
psutil==7.0.0
paho-mqtt